Fix FLIST_SAME_*() handling for the first entry
smb rsync may transmit these on the first entry because it doesn't use any
kind of first-file sentinel when comparing.  If we see them, it's clear that
we just need to zero the field out as smb rsync was comparing against a
variable in .bss.

Add a test for the obvious regression; device numbers and mode are going to
be harder or outright impossible to hit, they're mostly included here
because we know what to do if it does crop up.