Fixed itemize transfer direction bug (< vs >).
Fixed bug where mkpathat() was being called with a read-only buffer.

Fixed bug where -k and -L were not being sent to remote receiver
and -K was being sent to remote sender.

Add %I to log-format when sending to remote receiver when either -ii
or -ivv is given.

Ensure deleted files are logged with "*deleting" when -i --delete is given.

Prevent rsync_set_metadata() from changing things during dry run.

Fixed logic to detect whether or not the USER itemization flag
should be set (i.e., depends on --super and/or geteuid()).

Fixed invalid dry-run item logging with -L and -k for symlinks that
point to existing directories.

Fixed itemization bug that didn't correctly report file size changes.

Fixed bug that didn't correctly set the NEW and TRANSFER itemization
flags when the source file is regular and the dest file is not.

Fixed item logging issues arising from a various combinations
of -i, -v, and the the remoteness of the receiver.

Make -v and -vv output similar to rsync2/rsync3.

Added a test for itemize-changes.