Change the way find_hl() works.
We can't be sure that files will be transferred in the order they
appear in the flist, so we can't use the "first file" to know which
file is the original, and which are the additional hardlinks.

Instead we set the FLIST_NEED_HLINK flag on every hardlink and then
when a file is transferred, we remove the flag.

When searching for the leader of a group of hardlinks, we use the
first file we find that has already been transferred, and hardlink
the rest of the files to that file.