openrsync: fmap: start passing sizes to fmap_data()
In a near future change, we'll add an implementation of the fmap interface
that instead uses read/lseek to fetch the pertinent parts of a file into a
local buffer.  The size passed in to fmap_data() will be used to size our
buffer to make sure we have everything we need for contiguous access by the
caller.