openrsync: take greater care to avoid flagging EOF past requested size
We may try to eagerly overread in case a subsequent caller wants the next
chunk of the file, to try and minimize syscall overhead.  In some cases, we
could have expanded our buffer enough that we'd overread and flag the EOF,
even though the caller didn't even ask for the extra bits.