openrsync: don't truncate input missing EOL prematurely
fdgets() assumes that every file in the listing will be either newline or
NUL terminated, depending on the options used.  However, it misses that a
filename at EOF doesn't need to be terminated- we still need to pick it up.

Just break and return the current length once we hit EOF, instead of
truncating the listing.