openrsync: ensure that fdgets() properly terminates the buffer
The main problem here was introduced in
e95ae8741 (openrsync: don't truncate input missing EOL prematurely)
but the secondary problem that we'll actually read up to the full buffer
size was noticed while working on this patch.

Rewrite fdgets() to be a little more straightforward to read through; we now
read the single byte onto the stack, then interpret it and optionally append
it to the buffer.  fdgets() hasn't been able to read more than a single byte
at a time in a long time to avoid over-reading, so it seems better to avoid
that pretense entirely.