openrsync: use the local file position when fast-forwarding compressor
When we send a token for the receiver to use, we have to feed the same data
into the compressor to keep the stream in a consistent state.  Previously we
were using the token we sent to calculate an offset, but the token actually
has very little to do with the local file!  It's actually an index into the
receiver's block set, so it could very well extend beyond the end of our
copy.

Switch to using the current position in the file, which is what we should
have been using.  We still use the block information for the size of the
chunk, since the fact that we matched it implies that the size was actually
right- we'll just sanity check it against the remainder of the file.