openrsync: fmap: fallback to bufio if mmap() fails due to ENOMEM
There's no reason for this particular error to be fatal; the memory
requirements for bufio will be much lower, with the trade-off that it will
not necessarily be as fast to access.  Slower is better than not at all, so
fallback to bufio if we don't have the memory to mmap() it.