openrsync: fix progress output to go to stdout, don't send to server
smb rsync uses FCLIENT for this, which means it goes to stdout and *not* to
the log file, so we can't use LOG0.  Just write it directly to stdout.

While we're here, don't honor the flag if we happen to get it as the server
and stop sending it to the server if we're the client -- smb rsync won't.