openrsync: defer handling of hangup until necessary in the sender
When we observe a hangup from the remote end, drain the socket and flag the
iobuf as eof.  Defer erroring out until we need more data than we currently
have in the iobuf, just in case it would have been sufficient to finish out
the sender's state machine.

This fixes a race between the sender finishing up and the receiver
terminating, where the receiver can exit before the sender has read
everything from the socket on loaded systems.

This exact style of race is really unique to this configuration of
sender/receiver; the receiver drives the transfer progesrsion and thus has
a chance to finish up quicker.  If the client is the receiver instead, it
will ultimately block until the sender has finished closing out its state.