openrsync: fix last-component processing for --safe-links
dp will always be NULL after the loop, as we don't have any explicit breaks
that would put us there before we hit the end of the path.  Track the last
component and zap it if we have a trailing slash so that we can properly
handle a trailing .. component.