openrsync: fix accidental trailing space sensitivity
It was intended for trailing whitespace to be stripped later, but that part
appears to have been dropped in an intermediate diff.  What likely happened
was that it was intended to be moved when support for line continuations
was added so that we didn't strip internal whitespace, but it didn't make it
back in.

We add in some trailing whitespace in one of the tests arbitrarily to make
sure that this doesn't happen again.