openrsync: track rootfds for all flist entries, use O_RESOLVE_BENEATH
For non-directories that are named on the command line, we have no rootfd
and can't really enforce any particular constraints.

For directories named on the command line, we'll open a froot before we
start to recurse and refcount it: once for each path in the hierarchy.

O_RESOLVE_BENEATH is disabled for some command-line options that
specifically request symlinks be traversed.  These are not safe to use for,
e.g., backup applications.