tests: apply an -xattr find(1) filter to avoid files with newlines
This isn't portable upstream, but in the interest of getting a fix out the
door for our imminent problem it's fine -- this hasn't been tested again
upstream anyways, so we'll solve the portability issue at that time.

While we're here, make sure we're checking dirs and symlinks as well, and
apply the appropriate -h or -s flag in xattr_dump to get symlink xattrs and
not what they point at.