openrsync: escalate directories to writable for the transfer
Previously we'd only touch the permissions early if we were preserving
permissions, but that's not really what we want- the final permissions may
also not permit writing.  Upgrade permissions to user-writable beforehand
so that we can complete the update, they'll get downgraded later if -p was
specified or, apparently we can just leave it u+w if not.