openrsync: defer setting metadata until after closing the file
After some experimentation, samba (at least 4.16 and some other unknown
versions) seems to update the mtime on close(2) following a write(2).  The
problem is that it does so without considering whether a previous explicit
mtime set (e.g., via utimensat(2)) should clobber the update.  Our options
are to either fsync(2) everything before we attempt to set the metadata, or
just set them after close instead to get proper serialization of the two
changes.

fsync(2) on every single file is a lot of overhead, and trying to do it only
if we can detect an smb target is shadey at best.  Switch to setting it
after close(2), which is at least not a regression from a user switching
from smb rsync.