tests: avoid copying to a directory outside of destination root
For security reasons, we'll usually try not to allow the receiver write
outside of its transfer root to avoid, e.g., confused deputy style of
attacks.  Further testing on platforms with AT_RESOLVE_BENEATH logic
revealed this test causing some breakage, as it tries to write to
dst/../keepdir.

This is an easy fix that's mostly not ugly: move the symlink in the
destination to a path within the transfer root.  There's one spot where we
move it out so that we can compare that the rest of the tree otherwise looks
relatively OK, but it's not really needed after that so we can just remove
it.