openrsync: straighten out server logging
If we're logging to syslog, that shouldn't prevent us from sending a log
message to the client via the pipe if we need to.  We may be logging early
in the daemon, in which case we should just ignore the pipe if multiplexing
isn't enabled yet.

Servers should also stop immediately after the logfile, since they won't
have a sensible stdout/stderr from our perspective.

Servers should also never log before the transfer.