openrsync: log items at the appropriate time
The conditional logging, it turns out, is only used for items that we don't
need to transfer.  Transfers will always log either before or after the
transfer and shouldn't consult the iflags further.

The former log_item_impl() seems to have mostly been capturing some complex
logic that we didn't need, so it's GC'ed entirely since we pass the tests
without doing any of that.