openrsync: refactor out cfg_parse() calls
Encoding the logic for consulting the default config into daemon_cfg_parse()
is easier than making sure role->cfg_file is set to the default value, which
has the downside of not really allowing us to tell if we just happened to
have been explicitly provided the default config file as our config file, or
if we defaulted on our own.

That downside doesn't really matter at the moment, but it could be useful to
know later.