openrsync: fix --extended-attributes for non-files on macOS
We have a few issues here that we're addressing:

1.) The AppleDouble files get created as whatever type they ultimately apply
  to, but they should always be files

2.) platform_finish_transfer() isn't called for non-files, which is where
  the Apple platform bits know to merge an AppleDouble file in if present

3.) The merging bits did not previously think about having to apply xattrs
  to anything but files, but open(2) needs different flags.

The --links -> O_NOFOLLOW implication, I suspect, was a misguided attempt to
do what O_SYMLINK is meant to do.  If we're doing --no-links, then we would
not have a symlink to follow in the first place.

While we're here, make sure we test xattrs on both directories and symlinks.