openrsync: truncate the file at end of transfer for --inplace
Files shorter than the original file in the destination would previously not
cause a truncation, so we'd end up with stale data at the end of the file
that did not match the source.  We could use O_TRUNC, but for some reason
smb rsync decided to opt for truncating it at the end instead- we follow
suit to preserve semantics if a transfer is interrupted.