openrsync: avoid falling off the end of the arg list while logging
Previous changes straightened out our logging so that we don't miss logging
things (we may log to a few different places depending on --log-file), but
overlooked that the va_list advances as it's consumed by something that's
processed the format string.

Tap out a copy of the va_list for the first two times that we may log; the
daemon can hit the first two logging paths, the server can only hit the
second and third, and the client can only hit the first and third.  Instead
of trying to unwind the logic here, just tap out copies if there 'could' be
another output.

This fixes some weird failures in logging that we've seen, like a previous
PAC failure that didn't make any sense and other failures that Greg has
observed with ASAN enabled.  I haven't yet found a good way to reliably
induce poor behavior in a test.