openrsync: downloader: streamline our buf_copy() implementation
We need identical handling for all of the buffers that we write out to the
file, so it makes more sense to implement a state machine here.  We'll later
want to track our position so that we can examine segments in terms of the
hole/block size to try and determine if we can create a hole.