openrsync: downloader: fix our --sparse logic
Process buffers that we're in the process of writing out in terms of the
hole size we infer from the file we're writing, either via fpathconf() or
by inspection of the block size.

This also makes our test case a little more interesting by creating some
holes interspersed with data, so that we can be sure that a surprise
non-hole segment didn't stop us from tracking holes later in the file.