openrsync: sender: try a little harder to flush out our writebuf
We need to account for short writes and other errors here.  While we may not
be able to send any further errors over the socket, we might have a local
log configured if it's, e.g., an rsync daemon.  It may be the case that we
can't really get anything out, in which case we'll just abandon ship since
we're in an error path anyways.