openrsync: clear iflags in the redo phase
When the receiver machinery drives us into the redo phase, we should clear
iflags before doing any pre_*() checks in case we just entered the redo
phase.  Many of our implementations may only OR new flags in because some
paths through the function will set them differently, so we do want to clear
these to avoid unexpected combinations later.

The fact that we have unlinked the file locally is an implementation detail
that may or may not be technically incorrect, but it achieves the correct
effect.  We should revisit that, but for the time being we should get the
iflags correct.