openrsync: downloader: sanity check our mapped offsets
openrsync is structured in such a way that there's a lot of possibility for
a file to become stale between the uploader and the downloader on the
receiver side.  This isn't something we necessarily want to fix: the odds of
racing against some other process mutating it are low, and fixing it may not
scale very well for large transfers where the uploader can work ahead of the
downloader by quite a bit.

Currently we'd hit an assertion in fmap_data() when the token offset lies
outside of our range.  This isn't caught by the previous tok/blksz checks
because blksz comes from the sender, which is operating on information we
provided in the uploader.  We have to check both to be technically correct.

Note that we can't always provide this earlier sanity check; if we're using
a BUFIO fmap, then we don't know the size in advance.  Thus, we pull the
size check out into an fmap_access_valid() that will effectively be a NOP
if BUFIO is in use.  Skipping it is fine since we'll simulate a SIGBUS and
return to the trap if we're not allowed to read to the extent that we tried
to.

While we're here, fix the unmapped file error to be a warning instead; this
isn't an I/O error unless we hit it in a redo attempt, in which case it
gets bubbled back up as another hash mismatch and flagged as an I/O error.