openrsync: address another race against truncation in the receiver
pread() here can fail either with an error or a short read if the file was
truncated between collecting information about it in pre_file() and sending
block information.  For the latter case, we don't have to fail the transfer,
we just proceed as if --whole-file were specified because something weird is
happening.